Rename ResultContainer to GridResultContainer and move to folder
