+++ psdm_analysis/plots/results/flex_plot.py
@@ -5,7 +5,7 @@ from matplotlib.axes import Axes
 
 from psdm_analysis.models.enums import SystemParticipantsEnum
 from psdm_analysis.models.result.flex_option import FlexOptionResult
-from psdm_analysis.models.result.participant.participants_res_container import (
+from psdm_analysis.models.result.container.participants import (
     ParticipantsResultContainer,
 )
 from psdm_analysis.models.result.power import PQResult
